Fix various warnings with the Intel C compiler.
EfiBind.h:
* Adding ignores for certain warnings with the Intel C compiler.
  (Partially merged from MdePkg.)

EfiPxe.h:
* Make sure PXE_NO_UINT64_SUPPORT is defined

Decompress.c:
* Removed 'unreachable' return statement.

EdkCompatibilityPkg/**: (all other files)
* Remove trailing comma from last value in enum.
* Include <EfiBind.h> instead of "EfiBind.h", or
  for some reason __STDC_VERSION__ is not defined.
* Introduce more explicit function call ordering by
  storing results in local variables before using
  in a calculation.
* Add some additional casting for BOOLEAN & UINT8
  following calculations which change the type to 'int'.

git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@5123 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/EdkCompatibilityPkg/Foundation/Library/EfiCommonLib/Math.c
@@ -167,7 +167,7 @@ Returns:
     Operand >>= 1;
     Bitpos++;
   }
-  return (Bitpos - 1);
+  return (UINT8)(Bitpos - 1);
 
 }
